+++ ui/src/SubHeader/Calendar/Calendar.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import '../Styles/Colors.scss';
+@import '../../Styles/Colors';
 
 .viewing-calendar  {
     width: 364px;
