+++ ui/src/SubHeader/Calendar/Calendar.tsx
@@ -20,7 +20,7 @@ import CalendarInputLabel from './CalendarInputLabel';
 import DatePicker from 'react-datepicker';
 import ReactDatePicker from 'react-datepicker';
 import React, {createRef, useEffect, useState} from 'react';
-import AssignmentClient from '../Services/Api/AssignmentClient';
+import AssignmentClient from '../../Services/Api/AssignmentClient';
 import moment from 'moment';
 import {useRecoilState, useRecoilValue} from 'recoil';
 import {ViewingDateState} from 'State/ViewingDateState';
