+++ ui/src/SubHeader/SortingAndFiltering/FilterOrSortBy.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import '../Styles/Colors';
+@import '../../Styles/Colors';
 
 .dropdownContainer {
   display: flex;
