+++ ui/src/SubHeader/SortingAndFiltering/ProductLocationFilter/ProductLocationFilter.test.tsx
@@ -16,16 +16,16 @@
  */
 
 import React from 'react';
-import {renderWithRecoil} from '../../Utils/TestUtils';
+import {renderWithRecoil} from '../../../Utils/TestUtils';
 import {screen, waitFor} from '@testing-library/react';
-import TestData from '../../Utils/TestData';
+import TestData from '../../../Utils/TestData';
 import {LocalStorageFilters} from '../FilterLibraries';
-import {ModalContents, ModalContentsState} from '../../State/ModalContentsState';
-import {RecoilObserver} from '../../Utils/RecoilObserver';
-import {localStorageEventListenerKey} from '../../Hooks/useOnStorageChange/useOnStorageChange';
-import MyTagsForm from '../../Tags/MyTagsForm';
+import {ModalContents, ModalContentsState} from '../../../State/ModalContentsState';
+import {RecoilObserver} from '../../../Utils/RecoilObserver';
+import {localStorageEventListenerKey} from '../../../Hooks/useOnStorageChange/useOnStorageChange';
+import MyTagsForm from '../../../Tags/MyTagsForm';
 import ProductLocationFilter from './ProductLocationFilter';
-import {LocationsState} from '../../State/LocationsState';
+import {LocationsState} from '../../../State/LocationsState';
 
 describe('Product Location Filter', () => {
     let modalContent: ModalContents | null;
