+++ ui/src/SubHeader/SubHeader.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import '../Styles/Colors.scss';
+@import '../Styles/Colors';
 
 .newSpaceSelectionContainer {
   display: flex;
