+++ ui/src/Tags/TagsModalContent.tsx
@@ -25,7 +25,7 @@ import EditTagRow from '../ModalFormComponents/EditTagRow';
 import AddNewTagRow from '../ModalFormComponents/AddNewTagRow';
 import {INACTIVE_EDIT_STATE_INDEX} from './MyTagsForm';
 import {TagClient} from '../Types/TagClient';
-import {FilterType} from '../SortingAndFiltering/FilterLibraries';
+import {FilterType} from '../SubHeader/SortingAndFiltering/FilterLibraries';
 import {useRecoilValue} from 'recoil';
 import {CurrentSpaceState} from '../State/CurrentSpaceState';
 import {Tag} from '../Types/Tag';
